Cleaning of code
I have read through the code uploaded by Bun-san and have cleaned it up
so that it is far easier to read.

I have also integrated Eclipse for C/C++ with the git and the scenario
folder so that writing code is made somewhat simpler. The complete
instructions will be placed in the repository Wiki as soon as possible.
It is not a simple movement.

So that Eclipse does not complain about libraries and actually helps to
find the respective header files, the includes have been changed to
search via the LD_LIBRARY_PATH instead of the local directory.
Fortunately the waf building tools do not suffer from immediate issues
with these changes.

Cheers.