Files to make it easier to run ns3 scenarios
In the random/ directory there is a random number generator based on the
geometric distribution to generate the size of content.

In the ndn-campus directory there is a bash script run-sim.sh that passes
arguments to the scenarios, making it easier to run them from the command line.
For run-sim.sh to work, the ns3 scenarios must at least implement the
following command line options:

--clients=NUM     Number of consumers (clients) in the scenario
--contentsize=NUM Size of the content in bytes for the scenario
--networks=NUM    Number of networks to be produced in the scenario
--producers=NUM   Number of producers (servers) in the scenario

Cheers.