Printing out location using MobilityModel
The print shows the initial position of an assigned node. I am guessing that
you can keep track of it when certain events happen within the code. For now
it just prints the initial position.

I had to include iostream for simple printing. For other printing methods,
this code will have to be revised.